Mini_Project/requirements.txt
django
python-dotenv
mysqlclient

numpy
pillow
torch
facenet-pytorch